メールを実際に作成し送る部分と、メールの内容を作成する部分を分離した。
